#1 - Added read method in Transaction class
The commit introduces a new 'read' function in the Transaction class, which facilitates fetching record(s) from the session based on their id(s). To correspond with this change, the read method in models/base.py was modified to use this new transaction read method. Additionally, the test for reading multiple records was implemented in test_multi_user.py file.