#1 - Add 'all' method and test, update db settings
In this commit, the 'all' method was added to the BaseModel class to enable fetching all records from a table. A corresponding test was also added in tests/test_multi_user.py module. The echo setting in the database configuration file was also updated to False to suppress SQL echo.