#1 - Add tests for record counts and subject append
Added new assertions to test the record counts of the Course and Subject classes. Also, introduced a new test to verify the correct functioning of the append method when adding subjects to a course. This involved creating new subjects and appending them to a course, then confirming their correct creation and association.