#1 - Add deep read functionality to Student model
A new method has been added to the Student model that allows for a "deep read" operation. This deep read retrieves the student record from the database, and, if the "deep" argument is set to True, it also fetches the associated Registration records for that student. A corresponding test case was added to verify the proper functioning of this new method.